requirements.txt
Django==3.1.2
djangorestframework